Moving NREL to NLR API keys and naming throughout
