+++ docs/resource/meteosat_prime_meridian_v4_api.md
@@ -1,7 +1,7 @@
 (solar_resource:msg_v4_api)=
 # Solar Resource: Meteosat Prime Meridian PSM v4
 
-There are two datasets that use the [NSRDB Meteosat Prime Meridian PSM v4 API](https://developer.nrel.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-download/) calls:
+There are two datasets that use the [NSRDB Meteosat Prime Meridian PSM v4 API](https://developer.nlr.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-download/) calls:
 - "MeteosatPrimeMeridianSolarAPI"
 - "MeteosatPrimeMeridianTMYSolarAPI"
     - supports solar resource data for typical meteorological year (TMY), typical global horizontal irradiance year (TGY), and typical direct normal irradiance year (TDY)
@@ -10,8 +10,8 @@ These datasets allow for resource data to be downloaded for **locations** within
 
 | Model      | Temporal resolution | Spatial resolution | Years covered | Regions | Website |
 | :--------- | :---------------: | :---------------: | :---------------: | :---------------: | :---------------: |
-| `MeteosatPrimeMeridianSolarAPI`  | 15, 30, 60 min  | 4 km | 2005-2022  | Africa and Europe | [Meteosat Prime Meridian](https://developer.nrel.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-download/) |
-| `MeteosatPrimeMeridianTMYSolarAPI`  | 60 min  | 4 km | 2022 or 2014, for tmy, tdy and tgy  | Africa and Europe |  [Meteosat Prime Meridian TMY](https://developer.nrel.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-tmy-download/), [Meteosat Prime Meridian TGY](https://developer.nrel.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-tgy-download/), and [Meteosat Prime Meridian TDY](https://developer.nrel.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-tdy-download/)  |
+| `MeteosatPrimeMeridianSolarAPI`  | 15, 30, 60 min  | 4 km | 2005-2022  | Africa and Europe | [Meteosat Prime Meridian](https://developer.nlr.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-download/) |
+| `MeteosatPrimeMeridianTMYSolarAPI`  | 60 min  | 4 km | 2022 or 2014, for tmy, tdy and tgy  | Africa and Europe |  [Meteosat Prime Meridian TMY](https://developer.nlr.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-tmy-download/), [Meteosat Prime Meridian TGY](https://developer.nlr.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-tgy-download/), and [Meteosat Prime Meridian TDY](https://developer.nlr.gov/docs/solar/nsrdb/nsrdb-msg-v1-0-0-tdy-download/)  |
 
 
 ```{note}
