+++ h2integrate/converters/solar/test/test_pysam_with_atb_costs.py
@@ -5,7 +5,7 @@ from pytest import fixture
 from h2integrate.converters.solar.solar_pysam import PYSAMSolarPlantPerformanceModel
 from h2integrate.converters.solar.atb_res_com_pv_cost import ATBResComPVCostModel
 from h2integrate.converters.solar.atb_utility_pv_cost import ATBUtilityPVCostModel
-from h2integrate.resource.solar.nrel_developer_goes_api_models import GOESAggregatedSolarAPI
+from h2integrate.resource.solar.nlr_developer_goes_api_models import GOESAggregatedSolarAPI
 
 
 @fixture
