+++ h2integrate/converters/wind/test/test_floris_wind.py
@@ -9,7 +9,7 @@ from h2integrate import ROOT_DIR, H2I_LIBRARY_DIR
 from h2integrate.core.utilities import load_yaml
 from h2integrate.converters.wind.floris import FlorisWindPlantPerformanceModel
 from h2integrate.resource.wind.openmeteo_wind import OpenMeteoHistoricalWindResource
-from h2integrate.resource.wind.nrel_developer_wtk_api import WTKNRELDeveloperAPIWindResource
+from h2integrate.resource.wind.nlr_developer_wtk_api import WTKNRELDeveloperAPIWindResource
 
 
 @fixture
