+++ h2integrate/resource/solar/nlr_developer_api_base.py
@@ -1,62 +1,17 @@
 import urllib.parse
-from pathlib import Path
 
 import pandas as pd
-from attrs import field, define
 
-from h2integrate.core.validators import range_val
-from h2integrate.resource.resource_base import ResourceBaseAPIConfig
-from h2integrate.resource.wind.wind_resource_base import WindResourceBaseAPIModel
-from h2integrate.resource.utilities.nrel_developer_api_keys import (
+from h2integrate.resource.solar.solar_resource_base import SolarResourceBaseAPIModel
+from h2integrate.resource.utilities.nlr_developer_api_keys import (
     get_nrel_developer_api_key,
     get_nrel_developer_api_email,
 )
 
 
-@define(kw_only=True)
-class WTKNRELDeveloperAPIConfig(ResourceBaseAPIConfig):
-    """Configuration class to download wind resource data from
-    `Wind Toolkit Data V2 <https://developer.nrel.gov/docs/wind/wind-toolkit/wtk-download/>`_.
-
-    Args:
-        resource_year (int): Year to use for resource data.
-            Must been between 2007 and 2014 (inclusive).
-        resource_data (dict | object, optional): Dictionary of user-input resource data.
-            Defaults to an empty dictionary.
-        resource_dir (str | Path, optional): Folder to save resource files to or
-            load resource files from. Defaults to "".
-        resource_filename (str, optional): Filename to save resource data to or load
-            resource data from. Defaults to None.
-
-    Attributes:
-        dataset_desc (str): description of the dataset, used in file naming.
-            For this dataset, the `dataset_desc` is "wtk_v2".
-        resource_type (str): type of resource data downloaded, used in folder naming.
-            For this dataset, the `resource_type` is "wind".
-        valid_intervals (list[int]): time interval(s) in minutes that resource data can be
-            downloaded in. For this dataset, `valid_intervals` are 5, 15, 30, and 60 minutes.
-
-    """
-
-    resource_year: int = field(converter=int, validator=range_val(2007, 2014))
-    dataset_desc: str = "wtk_v2"
-    resource_type: str = "wind"
-    valid_intervals: list[int] = field(factory=lambda: [5, 15, 30, 60])
-    resource_data: dict | object = field(default={})
-    resource_filename: Path | str = field(default="")
-    resource_dir: Path | str | None = field(default=None)
-
-
-class WTKNRELDeveloperAPIWindResource(WindResourceBaseAPIModel):
+class NRELDeveloperAPISolarResourceBase(SolarResourceBaseAPIModel):
     def setup(self):
-        # create the input dictionary for WTKNRELDeveloperAPIConfig
-        resource_specs = self.helper_setup_method()
-
-        # create the resource config
-        self.config = WTKNRELDeveloperAPIConfig.from_dict(
-            resource_specs,
-            additional_cls_name=self.__class__.__name__,
-        )
+        super().setup()
 
         # set UTC variable depending on timezone, used for filenaming
         self.utc = False
@@ -73,19 +28,18 @@ class WTKNRELDeveloperAPIWindResource(WindResourceBaseAPIModel):
             else:
                 self.interval = int(min(self.config.valid_intervals))
 
-        super().setup()
-
         # get the data dictionary
         data = self.get_data(self.config.latitude, self.config.longitude)
 
         self.resource_data = data
-
-        # add resource data dictionary as an output
-        self.add_discrete_output("wind_resource_data", val=data, desc="Dict of wind resource data")
+        # add resource data dictionary as an out
+        self.add_discrete_output(
+            "solar_resource_data", val=data, desc="Dict of solar resource data"
+        )
 
     def create_filename(self, latitude, longitude):
         """Create default filename to save downloaded data to. Filename is formatted as
-        "{latitude}_{longitude}_{resource_year}_wtk_v2_{interval}min_{tz_desc}_tz.csv"
+        "{latitude}_{longitude}_{resource_year}_{config.dataset_desc}_{interval}min_{tz_desc}_tz.csv"
         where "tz_desc" is "utc" if the timezone is zero, or "local" otherwise.
 
         Args:
@@ -125,15 +79,14 @@ class WTKNRELDeveloperAPIWindResource(WindResourceBaseAPIModel):
             "api_key": get_nrel_developer_api_key(),
             "email": get_nrel_developer_api_email(),
         }
-        base_url = "https://developer.nrel.gov/api/wind-toolkit/v2/wind/wtk-download.csv?"
-        url = base_url + urllib.parse.urlencode(input_data, True)
+        url = self.base_url + urllib.parse.urlencode(input_data, True)
         return url
 
     def load_data(self, fpath):
         """Load data from a file and format as a dictionary that:
 
-        1) follows naming convention described in WindResourceBaseAPIModel.
-        2) is converted to standardized units described in WindResourceBaseAPIModel.
+        1) follows naming convention described in SolarResourceBaseAPIModel.
+        2) is converted to standardized units described in SolarResourceBaseAPIModel.
 
         This method does the following steps:
 
@@ -148,39 +101,43 @@ class WTKNRELDeveloperAPIWindResource(WindResourceBaseAPIModel):
 
         Returns:
             dict: dictionary of data in standardized units and naming convention.
-            Time information is found in the 'time' key.
         """
-
-        data = pd.read_csv(fpath, header=1)
-        header = pd.read_csv(fpath, nrows=1, header=None).values[0]
-        header_keys = header[0 : len(header) : 2]
-        header_vals = header[1 : len(header) : 2]
+        data = pd.read_csv(fpath, header=2)
+        header = pd.read_csv(fpath, nrows=2, header=None)
+        header_keys = header.iloc[0].to_list()
+        header_vals = header.iloc[1].to_list()
         header_dict = dict(zip(header_keys, header_vals))
+
+        time_cols = ["Year", "Month", "Day", "Hour", "Minute"]
+        data_cols = [c for c in data.columns.to_list() if c not in time_cols]
+        colnames_to_units = {
+            c: header_dict[f"{c} Units"] for c in data_cols if f"{c} Units" in header_dict
+        }
+        # data_missing_units = [c for c in data_cols if f"{c} Units" not in header_dict]
+
+        colname_mapper = {c: f"{c} ({v})" for c, v in colnames_to_units.items()}
+
         site_data = {
-            "site_id": header_dict["SiteID"],
-            "site_tz": header_dict["Site Timezone"],
-            "data_tz": header_dict["Data Timezone"],
-            "site_lat": header_dict["Latitude"],
-            "site_lon": header_dict["Longitude"],
+            "id": int(header_dict["Location ID"]),
+            "site_tz": float(header_dict["Local Time Zone"]),
+            "data_tz": float(header_dict["Time Zone"]),
+            "site_lat": float(header_dict["Latitude"]),
+            "site_lon": float(header_dict["Longitude"]),
+            "elevation": float(header_dict["Elevation"]),
             "filepath": str(fpath),
         }
 
         data = data.dropna(axis=1, how="all")
+        data = data.rename(columns=colname_mapper)  # add units to colnames
 
-        data, data_units = self.format_timeseries_data(data)
-        # make units for data in openmdao-compatible units
-        data_units = {
-            k: v.replace("%", "percent").replace("degrees", "deg").replace("hour", "h")
-            for k, v in data_units.items()
-        }
-        data_units_temp = {k: "degC" for k, v in data_units.items() if v == "C"}
-        data_units.update(data_units_temp)
-        # convert data to standardized units
+        # dont include data that doesn't have units
+        data_main_cols = time_cols + list(colname_mapper.values())
+        # make units for data in openMDAO-compatible units
+        data, data_units = self.format_timeseries_data(data[data_main_cols])
+        # convert units to standard units
         data, data_units = self.compare_units_and_correct(data, data_units)
 
-        # include site data with data
         data.update(site_data)
-
         return data
 
     def format_timeseries_data(self, data):
@@ -205,11 +162,19 @@ class WTKNRELDeveloperAPIWindResource(WindResourceBaseAPIModel):
         for c in data_cols_init:
             units = c.split("(")[-1].strip(")")
             new_c = c.replace("air", "").replace("at ", "")
-            new_c = new_c.replace(f"({units})", "").strip().replace(" ", "_").replace("__", "_")
+            new_c = (
+                new_c.replace(f"({units})", "").strip().replace(" ", "_").replace("__", "_").lower()
+            )
 
-            if "surface" in c:
-                new_c += "_0m"
-                new_c = new_c.replace("surface", "").replace("__", "").strip("_")
+            if units == "c":
+                units = units.upper()
+            if units == "w/m2":
+                units = "W/m**2"
+            if units == "nan" or units == "%":
+                units = "percent"
+            if units == "Degree" or units == "Degrees":
+                units = "deg"
+
             data_rename_mapper.update({c: new_c})
             data_units.update({new_c: units})
         data = data.rename(columns=data_rename_mapper)
