+++ h2integrate/resource/test/conftest.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from h2integrate.resource.utilities.nrel_developer_api_keys import set_nrel_key_dot_env
+from h2integrate.resource.utilities.nlr_developer_api_keys import set_nrel_key_dot_env
 
 from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
 
@@ -65,7 +65,7 @@ def pytest_sessionstart(session):
     os.environ["OPENMDAO_REPORTS"] = "none"
 
     # Set a dummy API key
-    os.environ["NREL_API_KEY"] = "a" * 40
+    os.environ["NLR_API_KEY"] = "a" * 40
     set_nrel_key_dot_env()
 
     # Set RESOURCE_DIR to None so pulls example files from default DIR
